size check

+++ runtime4/major_gc.c
@@ -230,7 +230,7 @@ Caml_inline void mark_stack_push(struct mark_stack* stk, value block,
   }
 
   if (Wosize_val(block) >= 5000000) {
-    fprintf(stderr, "mark_stack_push with block=0x%p: ridiculous size of %d\n", (void*) block, Wosize_val(block));
+    fprintf(stderr, "mark_stack_push with block=0x%p: ridiculous size of %d\n", (void*) block, (int) Wosize_val(block));
     caml_attach_gdb(Val_unit);
   }
